requirements.txt
alembic==1.13.1
blinker==1.8.2
cachelib==0.9.0
circuitbreaker==2.0.0
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
Faker==25.2.0
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Cors==4.0.1
Flask-DotEnv==0.1.2
Flask-HTTPAuth==4.8.0
Flask-Limiter==3.7.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
greenlet==3.0.3
gunicorn==22.0.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.12.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
marshmallow-sqlalchemy==1.0.0
mdurl==0.1.2
mysql-connector-python==8.4.0
ordered-set==4.1.0
packaging==24.0
psycopg2-binary==2.9.9
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
rich==13.7.1
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
Werkzeug==3.0.3
wrapt==1.16.0
